Allow placing signs on all blocks next to doors
Signs that were placed there already worked, you just needed something like WorldEdit to place the sign. Or you could place a sign there because the block you placed a sign on was a protectable type itself, such as a chest.

Now it just works just like the blocks above and below a door.